+++ src/options.ts
@@ -33,7 +33,7 @@ const defaultRetroarchConfig: RetroArchConfig = {
 const cdnBaseUrl = 'https://cdn.jsdelivr.net/gh'
 
 const coreRepo = 'arianrhodsandlot/retroarch-emscripten-build'
-const coreVersion = 'v1.16.0'
+const coreVersion = 'v1.17.0'
 const coreDirectory = 'retroarch'
 
 const shaderRepo = 'libretro/glsl-shaders'